Set up and document deployment strategy for jobs
The included `deploy-oozie-job` script makes it easy to upload a set of
job files from an analytics client to HDFS and deploy them as an Oozie
by handling the complexity of acting as the `analytics-product` user,
using the latest timestamped refinery deployment, ensuring that job
files are always put into the same HDFS location, providing a
streamlined testing method, and filtering logspam.

This patch also updates the `wikipediapreview_stats` job to use the
Product Analytics email address for alerts and implement the new
deployment strategy.

Bug: T261953
Bug: T267940
Change-Id: I791affa533b24d21b4fc1a775cc52e1228c869c8