movement_metrics: Fix whitespace PyHive error
By not starting a Hive SET command with a newline char,
see T334442 for additional details. This bug seems to have
been introduced when migrating from anaconda-wmf to
conda-analytics and using latest wmfdata. The better
fix may be to strip whitespace chars on wmfdata before
passing to PyHive, but this serves as a hotfix.

Bug: T334302
Bug: T332896
Change-Id: I323fa896efc8331b89074942d8a57fa7728c42b6